requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
supabase
python-dotenv